Flytt send inn knapp til steg-komponent la den erstatte neste-knapp
