+++ src/components/SøknadsSteg/3-Oppsummering/Oppsummering.less
@@ -14,8 +14,3 @@
         margin-bottom: 3rem;
     }
 }
-
-.sendInn {
-    text-align: center;
-    margin-top: 2rem;
-}
