+++ src/components/SøknadsSteg/3-Oppsummering/Oppsummering.tsx
@@ -4,26 +4,9 @@ import Oppsummeringsbolk from './Oppsummeringsbolk';
 import { visLabelOgSvar } from '../../../utils/visning';
 import { useApp } from '../../../context/AppContext';
 import { RouteEnum } from '../../../routing/Routes';
-import { ISøknad } from '../../../typer/søknad';
-import Miljø from '../../../Miljø';
-import { Hovedknapp } from 'nav-frontend-knapper';
-import { send } from 'process';
 
 const Oppsummering: React.FC = () => {
     const { søknad, axiosRequest } = useApp();
-    const visInnsendingsKnapp = Miljø().visInnsendingKnapp;
-
-    function sendInnSøknad() {
-        axiosRequest<string, ISøknad>({
-            url: '/api/kontrakt',
-            method: 'POST',
-            withCredentials: true,
-            påvirkerSystemLaster: true,
-            data: søknad,
-        })
-            .then(console.log)
-            .catch(console.log);
-    }
 
     return (
         <Steg tittel={'Oppsummering'} erSpørsmålBesvart={true}>
@@ -46,11 +29,6 @@ const Oppsummering: React.FC = () => {
                         );
                     })}
             </Oppsummeringsbolk>
-            {visInnsendingsKnapp && (
-                <Hovedknapp type={'standard'} className={'sendInn'} onClick={sendInnSøknad}>
-                    Send inn søknad
-                </Hovedknapp>
-            )}
         </Steg>
     );
 };
